Remove 'Copy' variant from `enm::Edit`
Enum edits are always wrapped in `edit::Edit` when derived, which provides
a `Copy` variant. The underlying `enm::Edit::Copy` was never used
and can be safely removed as long as we always used the wrapped version
in our manual implementations, which should be infrequent.

This also means that we no longer have to match on an unreachable
variant for derived edits.